refactor(games-list): extract gamepad-nav input handlers into a part file
Phase 3 / A start: split the first slice of `_SystemGamesListState` out of
the my_games_list.dart monolith using the in-repo part/extension pattern
(mirrors providers/neo_sync_provider.dart). Zero behaviour / public-API change.

Moved verbatim into new `my_games_list/gamepad_nav.dart` as
`extension _GamepadNav on _SystemGamesListState`:
- `_initializeGamepad` (input mappings)
- `_handleStartButton`, `_handleLeftBumper`, `_handleRightBumper`
- `_navigateUp` / `_navigateDown` / `_navigateLeft` / `_navigateRight`

All state stays on the host State (extensions can't declare fields). The one
`setState` call in the moved code routes through a new host `rebuild()` bridge
because `State.setState` is `@protected` and can't be called from an extension
(same idiom as neo_sync_provider's `notify()`); behaviourally identical.

Host: 6 insertions (part directive + rebuild bridge) / 157 deletions.
format clean, `flutter analyze` no issues project-wide, 207/207 tests.
